Python3 upgrade

+++ install.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 set -e # everything must succeed.
-if [ ! -d venv ]; then
-    virtualenv --python=`which python2` venv
-fi
+. mkvenv.sh
 source venv/bin/activate
-pip install -r requirements.txt
+if [ -e requirements.lock ]; then
+    # just delete the .lock file when you want to recreate it
+    pip install -r requirements.lock
+else
+    pip install -r requirements.txt
+    echo "locking..."
+    pip freeze > requirements.lock
+    echo "wrote 'requirements.lock'"
+fi
 NEW_RELIC_EXTENSIONS=false pip install --no-binary :all: newrelic==2.82.0.62
